+++ commons.txt
@@ -8,7 +8,10 @@ Whole contact handling (serializing and writing)
 Invitation process
 NotificationCenter
 
-# EditRelays.test.tsx
+# ErrorMessage.tsx
+# InputElementHelper.tsx
+# Relays.tsx
+# relays.ts
 # ModalForm
