+++ src/NostrAuthContext.tsx
@@ -3,7 +3,7 @@ import { getPublicKey } from "nostr-tools";
 import { hexToBytes } from "@noble/hashes/utils";
 import { useApis } from "./Apis";
 import { DEFAULT_RELAYS } from "./nostr";
-import { sanitizeRelays } from "./components/EditRelays";
+import { sanitizeRelays } from "./relays";
 
 type Context = {
   user: KeyPair | undefined;
